Refactor update user logic

+++ user/src/main/java/com/etsia/user/application/service/UserDeleteService.java
@@ -1,25 +1,28 @@
 package com.etsia.user.application.service;
 
+import com.etsia.common.infrastructure.entities.User;
 import com.etsia.user.domain.repository.UserRepository;
-import com.etsia.user.domain.service.UserDomainService;
-import com.etsia.user.infrastructure.exception.ResourceNotFoundException;
-import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
+import com.etsia.user.domain.service.UUserDomainService;
+import com.etsia.user.infrastructure.exception.UserNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
-@AllArgsConstructor
 public class UserDeleteService {
 
-    @Qualifier("uURepository")
+    //@Qualifier("uURepository")
     private final UserRepository userRepository;
+    private final UUserDomainService UUserDomainService;
     //private final UserDomainService userDomainService;
 
+    public UserDeleteService(UserRepository userRepository, UUserDomainService uUserDomainService) {
+        this.userRepository = userRepository;
+        UUserDomainService = uUserDomainService;
+    }
+
     public void execute(Integer id){
-        try {
-            userRepository.Delete(id);
-        } catch (Exception e) {
-            throw new ResourceNotFoundException("User Not Found", id);
+        if(!UUserDomainService.FindById(id).isPresent()){
+            throw new UserNotFoundException("User Not Found");
         }
+            userRepository.Delete(id);
     }
 }
